% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHCheckRequestStatus}
\alias{RTHCheckRequestStatus}
\title{Polling the extraction status.
On Demand extraction requests are executed as soon as possible.
However, There is no guarantee on the delivery time.
If the previous request returned a monitor URL, RTHCheckRequestStatus must be executed until it returns the result.}
\usage{
RTHCheckRequestStatus(location, path, overwrite = FALSE)
}
\arguments{
\item{location}{The monitor URL.}

\item{path}{Path to content to.}

\item{overwrite}{Will only overwrite existing path if TRUE.}
}
\description{
Polling the extraction status.
On Demand extraction requests are executed as soon as possible.
However, There is no guarantee on the delivery time.
If the previous request returned a monitor URL, RTHCheckRequestStatus must be executed until it returns the result.
}
