% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHExtractRaw}
\alias{RTHExtractRaw}
\title{Performs an on demand extraction returning the raw results as a stream if the response is available in a short amount of time,
otherwise the server accepted the extracting and response with a monitor URL.
In the later case, You must poll the extraction status with RTHCheckRequestStatus.}
\usage{
RTHExtractRaw(b, path, overwrite = FALSE)
}
\arguments{
\item{b}{JSON request body. See REST API Reference Tree for format.}

\item{path}{Path to content to.}

\item{overwrite}{Will only overwrite existing path if TRUE.}
}
\description{
The result format is the native/raw result from the underlying extractor (usually csv).
}
