% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHGetUserPackageDeliveries}
\alias{RTHGetUserPackageDeliveries}
\title{Get the user package delivery and save it to disk}
\usage{
RTHGetUserPackageDeliveries(
  packageDeliveryId,
  path,
  overwrite = FALSE,
  aws = FALSE
)
}
\arguments{
\item{packageDeliveryId}{User package delivery Id}

\item{path}{Path to content to.}

\item{overwrite}{Will only overwrite ex}
}
\description{
Get the user package delivery and save it to disk
}
