% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHGetValidContentFieldTypes}
\alias{RTHGetValidContentFieldTypes}
\title{Returns a list of valid ContentFieldTypes for the report template type.}
\usage{
RTHGetValidContentFieldTypes(
  reportTemplateTypes = c("TickHistoryTimeAndSales", "TickHistoryMarketDepth",
    "TickHistoryIntradaySummaries")
)
}
\arguments{
\item{reportTemplateTypes}{Available template type for RTH are "TickHistoryTimeAndSales","TickHistoryMarketDepth", and "TickHistoryIntradaySummaries".}
}
\description{
Returns a list of valid ContentFieldTypes for the report template type.
}
