% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHHistoricalSearch}
\alias{RTHHistoricalSearch}
\title{Search for historical instruments given an instrument identifier.
Return instruments may be currently active, or inactive 'historical only' instruments.}
\usage{
RTHHistoricalSearch(
  identifier,
  startDateTime,
  endDateTime,
  identifierType = NULL,
  resultsBy = NULL
)
}
\arguments{
\item{identifier}{Instrument identifier}

\item{startDateTime}{The range's start date and time. The format is yyyy-mm-ddThh:mm:ss.sssZ}

\item{endDateTime}{The range's end date and time. The format is yyyy-mm-ddThh:mm:ss.sssZ}

\item{identifierType}{The type of identifier. Supported types are Ric, Isin, Cusip, Sedol. Search will look for the identifier in all supported types when not specified.}

\item{resultsBy}{Determines what information is returned for each found RIC. By RIC: Returns information purely based on the RIC history. By Entity: Returns information based on the entity associated with the RIC on the end date of the Range. This will cause RIC rename history to be returned. Defaults to searching by RIC when not specified.}
}
\description{
Search for historical instruments given an instrument identifier.
Return instruments may be currently active, or inactive 'historical only' instruments.
}
