% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHLogin}
\alias{RTHLogin}
\title{Request authentication token}
\usage{
RTHLogin(uname, pword)
}
\arguments{
\item{uname}{DSS username}

\item{pword}{DSS password}
}
\value{
An authentication token that must be applied to all requests
}
\description{
Request authentication token
}
