% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHReferenceHistory}
\alias{RTHReferenceHistory}
\title{Retrieve FID reference history events for a set of RICs in a specified date range.
Returns Collection Of ReferenceHistoryResult}
\usage{
RTHReferenceHistory(ricList, startDateTime, endDateTime)
}
\arguments{
\item{ricList}{The RIC identifiers to return reference history for.}

\item{startDateTime}{The range's start date and time. The format is yyyy-mm-ddThh:mm:ss.sssZ}

\item{endDateTime}{The range's end date and time. The format is yyyy-mm-ddThh:mm:ss.sssZ}
}
\description{
Retrieve FID reference history events for a set of RICs in a specified date range.
Returns Collection Of ReferenceHistoryResult
}
