% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHUserInfo}
\alias{RTHUserInfo}
\title{Retrieves a single User information.}
\usage{
RTHUserInfo(uname)
}
\arguments{
\item{uname}{DSS username}
}
\value{
Return list of ID, Name, Phone, and Email
}
\description{
Retrieves a single User information.
}
