% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{RTHUserPackageDeliveriesByDateRange}
\alias{RTHUserPackageDeliveriesByDateRange}
\title{List all user package deliveries (data files) for one package.}
\usage{
RTHUserPackageDeliveriesByDateRange(subscriptionId, fromDate, toDate)
}
\value{
Returns the list of deliveries by date range
}
\description{
List all user package deliveries (data files) for one package.
}
