% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRTH.R
\name{TRTHGetUserPackageDeliveries}
\alias{TRTHGetUserPackageDeliveries}
\title{Get the user package delivery}
\usage{
TRTHGetUserPackageDeliveries(PackageDeliveryId)
}
\arguments{
\item{PackageDeliveryId}{User package delivery Id}
}
\value{
The package csv file
}
\description{
Get the user package delivery
}
