% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRTH.R
\name{TRTHUserInfo}
\alias{TRTHUserInfo}
\title{Retrieves a single User information.}
\usage{
TRTHUserInfo(uname)
}
\arguments{
\item{uname}{DSS username}
}
\value{
Return list of ID, Name, Phone, and Email
}
\description{
Retrieves a single User information.
}
