% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRTH.R
\name{TRTHUserPackageDeliveriesByPackageId}
\alias{TRTHUserPackageDeliveriesByPackageId}
\title{List all user package deliveries (data files) for one package.}
\usage{
TRTHUserPackageDeliveriesByPackageId(PackageId)
}
\arguments{
\item{PackageId}{User package Id. Usually from TRTHUserPackages()}
}
\value{
Return the list of user package delivery Id
}
\description{
List all user package deliveries (data files) for one package.
}
