% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRTH.R
\name{TRTHUserPackages}
\alias{TRTHUserPackages}
\title{Retrieve the list of all user packages, i.e. packages to which I am entitled (for all subscriptions)}
\usage{
TRTHUserPackages()
}
\value{
Return the list of user package Id, user package name and the corresponding subscription name
}
\description{
Retrieve the list of all user packages, i.e. packages to which I am entitled (for all subscriptions)
}
